debt - extract classes into separate files for some parts
