+++ src/vs/workbench/browser/parts/sidebar/sidebarPart.ts
@@ -4,11 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import 'vs/css!./media/sidebarpart';
-import { localize } from 'vs/nls';
+import 'vs/workbench/browser/parts/sidebar/sidebarActions';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { CompositePart } from 'vs/workbench/browser/parts/compositePart';
 import { Viewlet, ViewletRegistry, Extensions as ViewletExtensions, ViewletDescriptor } from 'vs/workbench/browser/viewlet';
-import { Action2, registerAction2 } from 'vs/platform/actions/common/actions';
 import { IViewletService } from 'vs/workbench/services/viewlet/browser/viewlet';
 import { IWorkbenchLayoutService, Parts, Position as SideBarPosition } from 'vs/workbench/services/layout/browser/layoutService';
 import { IViewlet, SidebarFocusContext, ActiveViewletContext } from 'vs/workbench/common/viewlet';
@@ -16,8 +15,7 @@ import { IStorageService } from 'vs/platform/storage/common/storage';
 import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
-import { KeyMod, KeyCode } from 'vs/base/common/keyCodes';
-import { IInstantiationService, ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
+import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { Event, Emitter } from 'vs/base/common/event';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { contrastBorder } from 'vs/platform/theme/common/colorRegistry';
@@ -33,8 +31,6 @@ import { LayoutPriority } from 'vs/base/browser/ui/grid/grid';
 import { assertIsDefined } from 'vs/base/common/types';
 import { CompositeDragAndDropObserver } from 'vs/workbench/browser/dnd';
 import { IViewDescriptorService, ViewContainerLocation } from 'vs/workbench/common/views';
-import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
-import { CATEGORIES } from 'vs/workbench/common/actions';
 import { Gesture, EventType as GestureEventType } from 'vs/base/browser/touch';
 
 export class SidebarPart extends CompositePart<Viewlet> implements IViewletService {
@@ -306,41 +302,5 @@ export class SidebarPart extends CompositePart<Viewlet> implements IViewletServi
 		};
 	}
 }
-
-class FocusSideBarAction extends Action2 {
-
-	constructor() {
-		super({
-			id: 'workbench.action.focusSideBar',
-			title: { value: localize('focusSideBar', "Focus into Side Bar"), original: 'Focus into Side Bar' },
-			category: CATEGORIES.View,
-			f1: true,
-			keybinding: {
-				weight: KeybindingWeight.WorkbenchContrib,
-				when: null,
-				primary: KeyMod.CtrlCmd | KeyCode.KEY_0
-			}
-		});
-	}
-
-	async run(accessor: ServicesAccessor): Promise<void> {
-		const layoutService = accessor.get(IWorkbenchLayoutService);
-		const viewletService = accessor.get(IViewletService);
-
-		// Show side bar
-		if (!layoutService.isVisible(Parts.SIDEBAR_PART)) {
-			layoutService.setSideBarHidden(false);
-			return;
-		}
-
-		// Focus into active viewlet
-		const viewlet = viewletService.getActiveViewlet();
-		if (viewlet) {
-			viewlet.focus();
-		}
-	}
-}
-
-registerAction2(FocusSideBarAction);
 
 registerSingleton(IViewletService, SidebarPart);
